+++ lib/src/annotations.dart
@@ -1,7 +1,7 @@
 import 'package:meta/meta.dart';
 import 'package:meta/meta_meta.dart';
 
-@Target({TargetKind.method, TargetKind.function})
+@Target({TargetKind.method})
 final class _Mutation {
   const _Mutation();
 }